deps/v8/test/inspector/debugger/wasm-gc-in-debug-break-expected.txt

Tests GC within DebugBreak

Running test: test
Script wasm://wasm/38e28046 byte offset 51: Wasm opcode 0x20 (kExprLocalGet)
GC triggered
Debugger.resume
Hello World (v8://test/instantiate:11:36)
    at bar (v8://test/instantiate:11:36)
exports.main returned!
